prepare server side

+++ wallet/clear.php
@@ -1,23 +1,32 @@
 <?php
 include_once "../db.php";
-query("delete from users");
-query("delete from domains");
-query("delete from domain_keys");
-query("delete from offers");
-query("delete from coins");
+$access = get_required("access");
+$message = null;
 
-http_json_post($host_url . "create_coin", array(
-    "user_login" => "ss",
-    "user_password" => "123",
-    "coin_name" => "ftr",
-    "coin_code" => "FTR",
-));
+$message = $access == "admin" ? null : "access error";
 
-http_json_post($host_url . "create_coin", array(
-    "user_login" => "ww",
-    "user_password" => "123",
-    "coin_name" => "usd",
-    "coin_code" => "USD",
-));
+if ($message == null) {
 
-echo json_encode(true);
+    query("delete from users");
+    query("delete from domains");
+    query("delete from domain_keys");
+    query("delete from offers");
+    query("delete from coins");
+
+    $message = $message == null && http_json_post($host_url . "create_coin", array(
+        "user_login" => "ss",
+        "user_password" => "123",
+        "coin_name" => "ftr",
+        "coin_code" => "FTR",
+    )) ? null : "create ftr error";
+
+    $message = $message == null && http_json_post($host_url . "create_coin", array(
+        "user_login" => "ww",
+        "user_password" => "123",
+        "coin_name" => "usd",
+        "coin_code" => "USD",
+    )) ? null : "create usd error";
+
+}
+
+echo json_encode(array("message" => $message));
